Rename dynamic_community_detection to  community_detection_online
