+++ cpp/CMakeLists.txt
@@ -81,6 +81,6 @@ add_subdirectory(bipartite_matching_module)
 add_subdirectory(bridges_module)
 add_subdirectory(connectivity_module)
 add_subdirectory(cycles_module)
-add_subdirectory(dynamic_community_detection_module)
+add_subdirectory(community_detection_module)
 add_subdirectory(pagerank_module)
 add_subdirectory(uuid_module)
