+++ e2e/community_detection_online_test/test_get/test.yml
@@ -1,0 +1,36 @@
+query: >
+  CALL community_detection_online.get()
+  YIELD node, community_id
+  RETURN node.id AS node_id, community_id
+  ORDER BY node_id;
+output:
+  - node_id: 0
+    community_id: 1
+  - node_id: 1
+    community_id: 1
+  - node_id: 2
+    community_id: 1
+  - node_id: 3
+    community_id: 1
+  - node_id: 4
+    community_id: 2
+  - node_id: 5
+    community_id: 2
+  - node_id: 6
+    community_id: 2
+  - node_id: 7
+    community_id: 2
+  - node_id: 8
+    community_id: 2
+  - node_id: 9
+    community_id: 3
+  - node_id: 10
+    community_id: 3
+  - node_id: 11
+    community_id: 3
+  - node_id: 12
+    community_id: 3
+  - node_id: 13
+    community_id: 3
+  - node_id: 14
+    community_id: 3
