+++ e2e/community_detection_online_test/test_set_empty/test.yml
@@ -1,0 +1,6 @@
+query: >
+  CALL community_detection_online.set()
+  YIELD node, community_id
+  RETURN node.id AS node_id, community_id
+  ORDER BY node_id;
+output: []
