--- e2e/dynamic_community_detection_test/test_get_no_reset/test.yml
@@ -1,6 +1,0 @@
-query: >
-    CALL dynamic_community_detection.get()
-    YIELD node, community_id
-    RETURN COUNT(node) AS n_nodes;
-output:
-  - n_nodes: 5
