--- e2e/dynamic_community_detection_test/test_set_empty/test.yml
@@ -1,6 +1,0 @@
-query: >
-  CALL dynamic_community_detection.set()
-  YIELD node, community_id
-  RETURN node.id AS node_id, community_id
-  ORDER BY node_id;
-output: []
